Rohit | BAH-1975 | Upgrading OMRS from 2.1.6 to 2.4.2 (#84)
* BAH-1634  Upgrade OMRS version from 2.1.6 to 2.4.2 in Appointments Module (#80)

* upgraded library versions and added @PowerMockIgnore(javax.management.*) to test classes

* test case adjustments to accomodate  junit5

* ignored test cases for https://bahmni.atlassian.net/browse/BAH-1848

* modified appointments hibernate config un-ignored test cases from https://bahmni.atlassian.net/browse/BAH-1848

* modified scope to test for javax.mail artifact

* BAH-1649 - | Rohit | - Upgrades openmrs atomfeed version to 2.6.2-SNAPSHOT (#81)

* Bah 1649 - Upgrades openmrs atomfeed version to 2.6.2-SNAPSHOT (#83)

* BAH-1649 - | Rohit | - Upgrades openmrs atomfeed version to 2.6.2-SNAPSHOT

* BAH-1649 - | Rohit | - Adds version properties for version webservices and javaxmail

* OMRS-master-2.4 - | Bindu, Rohit | - Removes unused Ignore import

+++ omod/src/test/java/org/openmrs/module/appointments/web/controller/AppointmentServiceControllerIT.java
@@ -57,8 +57,8 @@ public class AppointmentServiceControllerIT extends BaseIntegrationTest {
         assertEquals("Cardiology Consultation", asResponse.get("name"));
         assertEquals("09:00:00", asResponse.get("startTime"));
         assertEquals("17:30:00", asResponse.get("endTime"));
-        assertEquals(30, asResponse.get("durationMins"));
-        assertEquals(30, asResponse.get("maxAppointmentsLimit"));
+        assertEquals(30, (int) asResponse.get("durationMins"));
+        assertEquals(30, (int) asResponse.get("maxAppointmentsLimit"));
         assertEquals("#00ff00", asResponse.get("color"));
         assertNotNull(asResponse.get("weeklyAvailability"));
         assertEquals(0, ((ArrayList)asResponse.get("weeklyAvailability")).size());
@@ -97,8 +97,8 @@ public class AppointmentServiceControllerIT extends BaseIntegrationTest {
         assertEquals("Cardiology Consultation", asResponse.get("name"));
         assertEquals("09:00:00", asResponse.get("startTime"));
         assertEquals("17:30:00", asResponse.get("endTime"));
-        assertEquals(30, asResponse.get("durationMins"));
-        assertEquals(30, asResponse.get("maxAppointmentsLimit"));
+        assertEquals(30, (int)asResponse.get("durationMins"));
+        assertEquals(30, (int)asResponse.get("maxAppointmentsLimit"));
         assertEquals("#0000ff", asResponse.get("color"));
         ArrayList weeklyAvailabilities = (ArrayList) asResponse.get("weeklyAvailability");
         assertNotNull(weeklyAvailabilities);
